# specs/002-avatar-video-pipeline/contracts/api.yaml
openapi: 3.0.3
info:
  title: Avatarium API
  description: Avatar video generation platform API
  version: 0.1.0
  contact:
    name: Avatarium
  license:
    name: MIT
servers:
  - url: http://localhost:8000
    description: Local development
  - url: https://api.avatarium.example.com
    description: Production
...
paths:
...
  /api/projects:
    get:
      tags: [Projects]
      operationId: listProjects
      summary: List user's projects
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: per_page
          in: query
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: Paginated list of projects
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Projects]
      operationId: createProject
      summary: Create a new video project
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
      responses:
        '201':
          description: Project created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
...
  /api/projects/{projectId}/generate:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
    post:
      tags: [Pipeline]
      operationId: startGeneration
      summary: Start the video generation pipeline
      description: |
        Triggers the full generation pipeline:
        1. Moderate scenario via Gemini
        2. Split into segments via Gemini
        3. Generate first image via fal.ai (Qwen Image)
        4. For each segment: generate video clip → extract last frame → repeat
        5. Concatenate all clips via FFmpeg
      security:
        - bearerAuth: []
      responses:
        '202':
          description: Pipeline started (acknowledgment only; poll GET /status for progress)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenerationAcceptedResponse'
        '400':
          description: Missing photos or scenario
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Pipeline already running
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/projects/{projectId}/status:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
    get:
      tags: [Pipeline]
      operationId: getPipelineStatus
      summary: Get current pipeline status
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Pipeline status with per-segment progress
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PipelineStatusResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ── Videos ────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ProjectId:
      name: projectId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'
  schemas:
...
    CreateProjectRequest:
      type: object
      required: [title, video_style]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        video_style:
          type: string
          enum: [animation, movie_like]
    ProjectResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        video_style:
          type: string
          enum: [animation, movie_like]
        status:
          type: string
          enum: [draft, submitted, moderating, splitting, reviewing, generating, concatenating, completed, rejected, failed]
        estimated_cost:
          type: number
          nullable: true
        actual_cost:
          type: number
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    ProjectListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ProjectResponse'
        page:
          type: integer
        per_page:
          type: integer
        total:
          type: integer
...
    GenerationAcceptedResponse:
      type: object
      description: Lightweight 202 acknowledgment for POST /generate. Poll GET /status for full pipeline progress.
      properties:
        status:
          type: string
          enum: [generating]
        project_id:
          type: string
          format: uuid
    PipelineStatusResponse:
      type: object
      properties:
        project_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [draft, submitted, moderating, splitting, reviewing, generating, concatenating, completed, rejected, failed]
        current_segment:
          type: integer
          nullable: true
        total_segments:
          type: integer
          nullable: true
        clips:
          type: array
          items:
            $ref: '#/components/schemas/ClipStatusResponse'
        estimated_cost:
          type: number
          nullable: true
    ClipStatusResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        sequence_number:
          type: integer
        status:
          type: string
          enum: [pending, generating, completed, failed]
        retry_count:
          type: integer
        error_message:
          type: string
          nullable: true

    # ── Videos ────────────────────────────────────────────
...
    ErrorResponse:
      type: object
      description: |
        Matches FastAPI's native HTTPException response format.
        Only the 'detail' field is used (no separate 'error' field).
      properties:
        detail:
          type: string
    ValidationErrorResponse:
      type: object
      properties:
        detail:
          type: array
          items:
            type: object
            properties:
              loc:
                type: array
                items:
                  oneOf:
                    - type: string
                    - type: integer
              msg:
                type: string
              type:
                type: string